REFACTOR : present -> like 변경
